Separate file for wizards #7788
review5963001
