+++ tryton.cfg
@@ -10,3 +10,4 @@ xml:
     user.xml
     sale.xml
     product.xml
+    wizard.xml
