+++ wizard.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0"?>
+
+<tryton>
+    <data>
+        <!--Import Orders Wizard-->
+        <record model="ir.action.wizard" id="wizard_import_orders">
+            <field name="name">Import Orders</field>
+            <field name="wiz_name">sale.channel.orders_import</field>
+            <field name="model">sale.channel</field>
+        </record>
+        <record model="ir.action.keyword" id="wizard_import_orders_keyword">
+            <field name="keyword">form_action</field>
+            <field name="model">sale.channel,-1</field>
+            <field name="action" ref="wizard_import_orders"/>
+        </record>
+        <record model="ir.ui.view" id="import_order_start_view_form">
+            <field name="model">sale.channel.orders_import.start</field>
+            <field name="type">form</field>
+            <field name="name">wizard_import_orders_start_form</field>
+        </record>
+    </data>
+</tryton>
